Day 14 part 2
Gave several wrong answers for this due to a stupid bug in the sequence calculation which happened to give the right answer for the test. Cue an hour of pen and paper work pulling my hair out trying to figure out what was wrong with my maths. Turns out nothing.